fix: Refactor Stateful Code
